First version of the RFT library
This is the very first "stable" version that is
working with membership management and log replication
of the raft configuration. It still does not send messages
to the routing manager, it prints logger messages at the point
those send messages should be implemented.

This version supports multi-threaded xApps running over the
RFT library. Support of the RFT calling RMR running on top of
SI95 is also implemented in this version.

xApp state replication is also implemented and allows both,
GLOBAL and PARTIAL state replication.
The type of replication is configured by using the RFT_REPLICA_SERVERS
environment variable, which is read by the RFT init function.

In GLOBAL replication, the RFT library replicates the xApp state
to all other xApps in the raft cluster configuration.
In PARTIAL replication, the RFT library replicates the xApp state
up to a maximun number of servers which is also defined by the
RFT_REPLICA_SERVERS environment variable.

The context-key replication feature is also implemented as part of
the xApp state replicatation.

This vesion allows that each server (leader or not) can set up its
own replica server. Each replica has an exact copy of the list of
servers array, thus each replica can take its decision safely based
on its local copy of the servers' list.

Co-authored-by: Alexandre Huff <alexandrehuff@gmail.com>