Changed default WG port from 4070 to random port
The default port of the workload generator (WG) was causing some
conflicts with xApps' ports. In some situations xApps
running on the host network were configured to used port 4070,
which conflicted with the default workload generator port.

This change implements the default port of workload generators to
pick a random port from 43000 to 43999. A specific port can still
be chosen by using the '-p' option.

Co-authored-by: Alexandre Huff <alexandrehuff@gmail.com>