Implemented a ring buffer to store rft messages
A linked list queue was used to store RFT messages sent
from the xApp application. Each enqueued message requires
memory allocation to the queue node structure. This results
in a considerable overhead when running the RFT in overloaded
environments. After dequeue each message the queue's node
resouces needs to be released.

This change replaces the linked list queue by a fixed-size FIFO
ring buffer. This removes the need to have memory allocation and
deallocation for every message enqueue and dequeued to/from the
ring buffer.

Co-authored-by: Alexandre Huff <alexandrehuff@gmail.com>