Implemented a specialized ring to store log entries
Log entries are stored in memory using an array list structure
which did not allow to take snapshots and to do log compaction
to save memory.

This change implements a specialized ring structure to store
all log entries, and allows that snapshotting and log compaction
can be implemented on top of it.

Snapshotting and log compaction are key concepts for the RFT library.
While snapshotting allows an xApp to send its current state to
another xApp, the log compaction avoids xApps running out of
memory due to unnecessary log entries.

Co-authored-by: Alexandre Huff <alexandrehuff@gmail.com>