Added Log Ratio feature to trigger xApp snapshots
The xApp snapshotting only trigger the snapshot function based on
the memory size (in bytes) of the stored log entries. This could
lead to fill up all the ring structure to store log entries in
case the memory threshold chose is too high.

This change fixes this problem by adding a Log Ratio value and
using the ratio value to check the number of log entries stored
in the log ring. When the number of stored entries reaches the
threshold imposed by the log ratio, then a new xApp snapshot is
triggered.