Implemented the role awareness feature
This change implements the role awareness feature to allow each xApp
replica learn its role that it is playing on a specific Managed
Equipment.
Three roles are available: RFT_PRIMARY, RFT_BACKUP, and RFT_NONE.
xApps can use those types of role to implement different behavior
accondingly to their responsibility.

The snapshot module also was updated to fit into this feature.
Contexts that are managed by the primary are passed to the xApp
state machine to be serialized.