Update RFT bootstrapping
RFT bootstrapping was done by setting up the RFT_BOOTSTRAP env var
to the name of the xApp that should become the first leader of the
cluster. But, when that replica (first leader) was restarted, then
it became the leader again without chech whether or not there were
other leader in the cluster.

This change updates the RFT bootstrapping by using Redis database
to store the current leader endpoint. When a replica is starting
it tries to update the leader endpoint to itself endpoint. This
only works if the key used to store that endpoint was not defined
yet in the database (SET command with NX flag). So, if the flag
was already set, then the new replica won't become the leader.
In such a case, the new replica fetches the current endpoint
of the leader from Redis and sends membership requests to that
enpoint in order to join the cluster.