Implement clear bootstrap endpoint from Redis
The bootstrap enpoint from Redes was not cleared when shutting down
the RFT cluster, thus requiring manual intervention on Redis database
to remove the corresponding bootstrap key.

This change implements this feature. Only the leader will remove the
key when the RFT cluster is shutting down. The leader waits twice the
election timeout before proceeding to remove the key. It only removes
the key if it was not changes by another leader (the key needs to point
to the leader being removed) to prevent that it will be removed in case
only the leader is being removed from the cluster.

Closes #1